+++ src/renderer/components/container/DatabaseNetworkGraph.tsx
@@ -4,8 +4,8 @@ import { useLocation } from 'react-router-dom';
 import { ipcRenderer } from 'electron';
 import {
   CartesianGrid,
-  Line,
-  LineChart,
+  Area,
+  AreaChart,
   ResponsiveContainer,
   Tooltip,
   XAxis,
@@ -78,7 +78,7 @@ export default function DatabaseNetworkGraph({ mode }: Props) {
       setHasGraphData(true);
       setGraph(
         <ResponsiveContainer width="100%" height="92%">
-          <LineChart data={normalizedData.reverse()}>
+          <AreaChart data={normalizedData.reverse()}>
             <CartesianGrid strokeDasharray="3" vertical={false} />
             <XAxis
               dataKey="timestamp"
@@ -90,8 +90,9 @@ export default function DatabaseNetworkGraph({ mode }: Props) {
             <YAxis
               allowDecimals={false}
               type="number"
-              domain={[0, (dataMax: number) => Math.ceil(dataMax / 250) * 250]}
+              domain={[0, (dataMax: number) => Math.ceil(dataMax) * 1.2]}
               tickFormatter={(el) => formatBytes(el)}
+              width={80}
             />
             <Tooltip
               content={(content) => {
@@ -99,24 +100,26 @@ export default function DatabaseNetworkGraph({ mode }: Props) {
               }}
             />
             {(mode === 'OUTBOUND' || mode === 'MIXED') && (
-              <Line
+              <Area
                 type="monotone"
                 dataKey="dbTrafficSent"
                 dot={false}
                 stroke="var(--green)"
+                fill="var(--light-green)"
                 strokeWidth={2}
               />
             )}
             {(mode === 'INBOUND' || mode === 'MIXED') && (
-              <Line
+              <Area
                 type="monotone"
                 dataKey="dbTrafficReceived"
                 dot={false}
                 stroke="var(--yellow)"
+                fill="var(--light-yellow)"
                 strokeWidth={2}
               />
             )}
-          </LineChart>
+          </AreaChart>
         </ResponsiveContainer>
       );
     }