+++ src/renderer/components/layout/EnvironmentArtifactsContainer.tsx
@@ -1,0 +1,11 @@
+import DefaultMotionDiv from '../base/DefaultMotionDiv';
+
+function EnvironmentArtifactsContainer() {
+  return (
+    <DefaultMotionDiv id="environment-artifacts-container">
+      <h3>Artefatos</h3>
+    </DefaultMotionDiv>
+  );
+}
+
+export default EnvironmentArtifactsContainer;