+++ src/renderer/components/layout/EnvironmentInsightsContainer.tsx
@@ -1,0 +1,11 @@
+import DefaultMotionDiv from '../base/DefaultMotionDiv';
+
+function EnvironmentInsightsContainer() {
+  return (
+    <DefaultMotionDiv id="environment-insights-container">
+      <h3>Insights</h3>
+    </DefaultMotionDiv>
+  );
+}
+
+export default EnvironmentInsightsContainer;