+++ src/renderer/components/layout/EnvironmentRuntimeStatsContainer.tsx
@@ -1,0 +1,11 @@
+import DefaultMotionDiv from '../base/DefaultMotionDiv';
+
+function EnvironmentRuntimeStatsContainer() {
+  return (
+    <DefaultMotionDiv id="runtime-stats-container">
+      <h3>Estatísticas De Runtime</h3>
+    </DefaultMotionDiv>
+  );
+}
+
+export default EnvironmentRuntimeStatsContainer;