+++ src/renderer/components/layout/EnvironmentServicesContainer.tsx
@@ -1,0 +1,11 @@
+import DefaultMotionDiv from '../base/DefaultMotionDiv';
+
+function EnvironmentServicesContainer() {
+  return (
+    <DefaultMotionDiv id="environment-services-container">
+      <h3>Histórico De Serviços</h3>
+    </DefaultMotionDiv>
+  );
+}
+
+export default EnvironmentServicesContainer;